Use volume's single api point to return label counts
