+++ src/Http/Controllers/Api/Volumes/FilterImageAnnotationsByLabelController.php
@@ -5,7 +5,6 @@ namespace Biigle\Modules\Largo\Http\Controllers\Api\Volumes;
 use Biigle\Volume;
 use Biigle\ImageAnnotation;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use Biigle\Http\Controllers\Api\Controller;
 
 class FilterImageAnnotationsByLabelController extends Controller
@@ -58,44 +57,5 @@ class FilterImageAnnotationsByLabelController extends Controller
             ->distinct()
             ->orderBy('image_annotations.id', 'desc')
             ->pluck('images.uuid', 'image_annotations.id');
-    }
-
-        /**
-     * Get all image labels and annotation count for a given volume
-     * 
-     * @api {get} 
-     * @apiGroup Volumes
-     * @apiName test
-     * @apiParam {Number} id The Volume ID
-     * @apiPermission user
-     * @apiDescription Returns a collection of project image labels, and annotation label counts
-     * 
-     * @apiSuccessExample {json} Success response:
-     * [{"id":1,
-     * "name":"a",
-     * "color":"f2617c",
-     * "parent_id":null,
-     * "label_tree_id":1,
-     * "source_id":null,
-     * "label_source_id":null,
-     * "uuid":"6d2e6061-9ed1-41df-92f0-4862d0d4b12e",
-     * "count":10}]
-     *
-     * @param int $id Volume ID
-     * @return \Illuminate\Database\Eloquent\Collection
-     */
-    public function getVolumeAnnotationLabels($id)
-    {
-        $volume = Volume::findOrFail($id);
-        $this->authorize('access', $volume);
-
-        return DB::table('labels')
-        ->join('image_annotation_labels', 'labels.id', '=', 'image_annotation_labels.label_id')
-        ->join('image_annotations', 'image_annotation_labels.annotation_id', '=', 'image_annotations.id')
-        ->join('images', 'image_annotations.image_id', '=', 'images.id')
-        ->where('images.volume_id','=',$id)
-        ->select('labels.*', DB::raw('COUNT(labels.id) as count'))
-        ->groupBy('labels.id')
-        ->get();
     }
 }
